update nn wrapper

+++ rtdetrv2_pytorch/src/zoo/rtdetr/rtdetr_criterion.py
@@ -82,7 +82,7 @@ class RTDETRCriterion(nn.Module):
                                     dtype=torch.int64, device=src_logits.device)
         target_classes[idx] = target_classes_o
 
-        target = F.one_hot(target_classes, num_classes=self.num_classes+1)[..., :-1]
+        target = F.one_hot(target_classes, num_classes=self.num_classes+1)[..., :-1].to(src_logits.dtype)
         loss = torchvision.ops.sigmoid_focal_loss(src_logits, target, self.alpha, self.gamma, reduction='none')
         loss = loss.mean(1).sum() * src_logits.shape[1] / num_boxes
